apps/sender/python/requirements.txt
psutil>=5.9.8
paho-mqtt>=2.1.0
